pyproject.toml
[tool.poetry]
name = "flexapytgbot"
version = "0.1.0"
description = "Telegram bot to display FlexUSD APY"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
python-telegram-bot = {version = "20.0a0", allow-prereleases = true}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
